% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inch_to_cm.R
\name{inch_to_cm}
\alias{inch_to_cm}
\title{inch to cm}
\usage{
inch_to_cm(x)
}
\arguments{
\item{x}{A numeric value}
}
\value{
A numeric value
}
\description{
Turning inch to cm.
}
\examples{
inch_to_cm(2)
}
\author{
Wendy Shue
}
