% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maternal_mortality.R
\docType{data}
\name{maternal_mortality}
\alias{maternal_mortality}
\title{Maternal Mortality data}
\format{
A data frame with 1028 rows and 4 variables:
\describe{
\item{iso}{country codes}
\item{country}{countries}
\item{X2000-X2019}{maternal mortality ratio from 2000-2019,respectively}
\item{FetalMal}{Indicator for fetal malformation}
}
}
\source{
https://data.worldbank.org/indicator/SH.STA.MMRT
}
\usage{
maternal_mortality
}
\description{
Data from a toxicology study blah blah blah
}
\keyword{datasets}
