% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long.R
\name{long}
\alias{long}
\title{wide data to long data}
\usage{
long(data)
}
\arguments{
\item{data}{A data frame in wide format}
}
\value{
A long data frame.
}
\description{
Pivoting data format from wide to long.
}
\examples{
long(maternal_mortality)
}
\author{
Wendy Shue
}
